pyproject.toml
[project]
name = "actions-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "python-dotenv>=1.1.0",
]
